pyproject.toml
[tool.poetry]
name = "arbitrum-gatherer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "arbitrum_gatherer"}]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.5.1"
pydantic = "^1.10.2"
requests = "^2.28.1"
web3 = "^5.31.1"
python-dotenv = "^0.21.0"
clickhouse-driver = "^0.2.4"
arrow = "^1.2.3"
black = "^22.10.0"
isort = "^5.10.1"
flake8 = "^5.0.4"
mypy = "^0.990"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
ipykernel = "^6.17.0"
jupyter = "^1.0.0"
notebook = "^6.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 99 # override black's default line-length
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | venv
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
# make it compatible with black
profile = "black" 

[tool.poetry.scripts]
my-project = "my_project.services.py:run"